fix(release-note): styles adjusts and move release notes to the correct page
